// src/utils/common/timezone.enum.ts
export enum SUPPORTED_TIME_ZONES {
  AMERICA_PEURTO_RICO = "America/Puerto_Rico",
  AMERICA_NEW_YORK = "America/New_York",
  AMERICA_CHICAGO = "America/Chicago",
  AMERICA_DENVER = "America/Denver",
  AMERICA_PHOENIX = "America/Phoenix",
  AMERICA_LOS_ANGELES = "America/Los_Angeles",
  AMERICA_ANCHORAGE = "America/Anchorage",
  AMERICA_HONOLULU = "Pacific/Honolulu"
}